Implements user registration with organization creation
Adds the ability for a user to register themselves and create a new organization in the process.
The registration flow now creates a user profile, an organization, and an org membership record linking them.

Removes the old UserService and FirestoreCollections.

The organization model now allows null values for twilio SID and Auth Token, to be configured later